split channels to separate files
